requirements.txt
fastapi
uvicorn
# SQLAlchemy
mysql-connector-python
python-dotenv
pytz
# async-sqlalchemy
asyncpg
databases
sqlalchemy[asyncio]
asyncmy
cloudinary
python-multipart
joblib
opencv-python
scikit-learn
matplotlib
pandas
tensorflow
nltk